Fix for complex anonymous function syntax
Correctly handles cases like
"    at Object.o.(anonymous function) [as throwsErr] (/home/vpupkin/src/demo/index.js:51:16)"

The drawback of this solution is that now each item in the `formats` array
can be either a single regexp or an array of regexps. On the other hand,
the regexps became more strict and robust.